refacto events_form & event_fields stimulus controllers
