--- app/javascript/controllers/events_forms_controller.js
@@ -1,8 +1,0 @@
-import { Controller } from "@hotwired/stimulus"
-
-// Connects to data-controller="event-form"
-export default class extends Controller {
-  onTrashClick(_) {
-    this.element.remove()
-  }
-}
